feat: 내정보 페이지에서 정보를 수정하면 홈에서 바뀐 정보가 출력되도록 수정함.

+++ src/main/kotlin/com/sessac/healthcare/presentation/userprofile/model/UserProfileUIModel.kt
@@ -3,6 +3,6 @@ package com.sessac.healthcare.presentation.userprofile.model
 data class UserProfileUIModel(
     val id: Long,
     var nickName: String,
-    var height: Int,
-    var weight: Int,
+    var height: Float,
+    var weight: Float,
 )
